Refactor contact form to contact.
